mark key macros for inlining
The java-truffle Mal adds support for 'inlining' macros, wherein
macros that are not sensitive to the environment and produce
code of size linear in the input can be expanded once and "inlined"
in place, avoiding re-expansion on each execution.

This provides an enormous performance benefit, especially for
self-hosted Mal, in which the key interpreter functions are
built around the cond macro.